restructured package per feedback
