--- src/student_success_tool/reporting/__init__.py
@@ -1,1 +1,0 @@
-from . import model_card
