--- src/student_success_tool/reporting/platforms/pdp/__init__.py
@@ -1,1 +1,0 @@
-from . import model_card
